@@ -7,6 +7,8 @@ from utils.airports import airport
 from utils.helpers import great_circle_distance, iso_to_timestamp
 from utils.types import Position
 
+from .query import QueryResult
+
 
 @dataclass
 class Mission:
@@ -16,6 +18,7 @@ class Mission:
     arrival: pd.Timestamp
     load_factor: float
     aircraft_type: str
+    flight_id: int | None = None
 
     @staticmethod
     def _airport_position(code: str) -> Position:
@@ -36,10 +39,11 @@ class Mission:
     def gc_distance(self) -> float:
         """Great circle distance between departure and arrival positions (m)."""
         return great_circle_distance(
-            self.origin_position.longitude,
             self.origin_position.latitude,
-            self.destination_position.longitude,
+            self.origin_position.longitude,
             self.destination_position.latitude,
+            self.destination_position.longitude,
+            degrees=True,
         )
 
     @classmethod
@@ -57,3 +61,15 @@ class Mission:
                 )
             )
         return result
+
+    @classmethod
+    def from_query_result(cls, qr: QueryResult, load_factor: float = 1.0) -> 'Mission':
+        return cls(
+            origin=qr.origin,
+            destination=qr.destination,
+            departure=qr.departure,
+            arrival=qr.arrival,
+            load_factor=load_factor,  # (real load factor not in QueryResult)
+            aircraft_type=qr.aircraft_type,
+            flight_id=qr.id,  # The schedule ID is unique across the database
+        )
